chore: rename Vue.js stores on pinia naming convention
refs: XRDDEV-1983
