+++ src/security-server/admin-service/ui/src/components/wizard/WizardPageGenerateCsr.vue
@@ -95,7 +95,7 @@ import Vue from 'vue';
 import { ValidationProvider, ValidationObserver } from 'vee-validate';
 import { mapActions, mapState } from 'pinia';
 import { useNotifications } from '@/store/modules/notifications';
-import { useCsr } from '@/store/modules/certificate-sign-request';
+import { useCsr } from '@/store/modules/certificateSignRequest';
 import { AxiosError } from 'axios';
 
 export default Vue.extend({
