+++ src/security-server/admin-service/ui/src/store/modules/initializeServer.ts
