+++ src/security-server/admin-service/ui/src/views/AddMember/MemberDetailsPage.vue
@@ -138,7 +138,7 @@ import { AddMemberWizardModes } from '@/global';
 import { validate } from 'vee-validate';
 import { mapActions, mapState, mapWritableState } from 'pinia';
 import { useNotifications } from '@/store/modules/notifications';
-import { useAddClient } from '@/store/modules/add-client';
+import { useAddClient } from '@/store/modules/addClient';
 import { useGeneral } from '@/store/modules/general';
 import { useUser } from '@/store/modules/user';
